Remove QR code sharing feature and unused SubjectiveReportForm component
