+++ pages/[teamSlug]/[seasonSlug]/[competitonSlug]/[reportId]/subjective.tsx
@@ -1,41 +1,122 @@
-import Container from "@/components/Container";
-import SubjectiveReportForm from "@/components/SubjectiveReportForm";
-import { Match, SubjectiveReportSubmissionType } from "@/lib/Types";
-import UrlResolver, { ResolvedUrlData } from "@/lib/UrlResolver";
-import { Analytics } from "@/lib/client/Analytics";
 import ClientApi from "@/lib/api/ClientApi";
 import { useCurrentSession } from "@/lib/client/useCurrentSession";
 import useInterval from "@/lib/client/useInterval";
-import { GetServerSideProps } from "next";
+import { Match, SubjectiveReport, SubjectiveReportSubmissionType } from "@/lib/Types";
 import { useRouter } from "next/router";
-import React from "react";
-import { useEffect, useState } from "react";
+import React, { createRef, FormEvent, useState } from "react";
+import { Analytics } from "@/lib/client/Analytics";
+import Loading from "@/components/Loading";
 
 const api = new ClientApi();
 
-export default function Subjective(props: ResolvedUrlData) {
+export default function SubjectiveReportForm(props: { match: Match, compId?: string, teamNumber?: number, compName?: string }) {
   const router = useRouter();
-  const { reportId: matchId } = router.query;
+  const { teamSlug, seasonSlug, competitonSlug } = router.query;
+  
+  const { match } = props;
+  
+  const session = useCurrentSession();
 
-  const [match, setMatch] = useState<Match | undefined>();
+  const submitButtonRef = createRef<HTMLButtonElement>();
 
-  useEffect(() => {
-    if (match || !matchId)
+  const [submitting, setSubmitting] = useState(false);
+
+  const [report, setReport] = useState<SubjectiveReport | undefined>(undefined);
+
+  function getReportFromForm(): SubjectiveReport {
+    return {
+      _id: undefined,
+      match: match._id as string,
+      matchNumber: match?.number ?? 0,
+      wholeMatchComment: (document.getElementById("comment-WholeMatch") as HTMLTextAreaElement)?.value ?? "",
+      robotComments: Object.fromEntries(
+        match?.blueAlliance.concat(match.redAlliance).map((team: number, index: number) => 
+          [match?.blueAlliance.concat(match.redAlliance)[index], (document.getElementById(`comment-${team}`) as HTMLTextAreaElement).value ?? ""])
+      ),
+      submitter: undefined,
+      submitted: SubjectiveReportSubmissionType.NotSubmitted
+    };
+  }
+
+  function updateReport() {
+    setReport(getReportFromForm());
+  }
+
+  async function submit(e: FormEvent<HTMLFormElement>) {
+    e.preventDefault();
+
+    setSubmitting(true);
+
+    let valid = [...(e.target as any)].filter((element: any) => element.value).length > 0;
+    if (!valid) {
+      if (submitButtonRef.current) {
+        const btn = submitButtonRef.current;
+
+        btn.classList.add("btn-error");
+        btn.innerText = "Please fill out at least one field";
+        setTimeout(() => {
+          btn.classList.remove("btn-error");
+          btn.innerText = "Submit";
+        }, 1500);
+      }
+
+      setSubmitting(false);
       return;
+    }
 
-    api.findMatchById(matchId as string).then(setMatch);
-  }, [match, matchId]);
+    api.submitSubjectiveReport(getReportFromForm(), teamSlug as string)
+      .then(() => {
+        const teamsWithComments = [...(e.target as any)].map((element: any, index: number) =>
+          ["Whole Match"].concat(match?.blueAlliance.concat(match.redAlliance).map((n) => n.toString()) ?? [])[index]
+        );
+    
+        Analytics.subjectiveReportSubmitted(teamsWithComments, props.teamNumber ?? -1, props.compName ?? "Unknown Competition", 
+          session.session.user?.name ?? "Unknown User");
+      })
+      .finally(() => {
+        if (location.href.includes("offline"))
+          location.href = `/offline/${props.compId}`;
+        else
+          location.href = `/${teamSlug}/${seasonSlug}/${competitonSlug}`;
+      });
+  }
 
-  if (!match)
-    return (
-      <Container requireAuthentication={true} hideMenu={false} title="Subjective Scouting">
-        <div>Loading...</div>
-      </Container>
-    );
-
-  return (
-    <Container requireAuthentication={true} hideMenu={false} title={`${match?.number} | Subjective Scouting`}>
-      <SubjectiveReportForm match={match} compId={props.competition?._id} teamNumber={props.team?.number} compName={props.competition?.name} />
-    </Container>
+  // We have to use router as a dependency because it is only populated after the first render (during hydration)
+  useInterval(() => api.checkInForSubjectiveReport(match._id as string), 5000, [router]);
+  
+  return (   
+    <div className="flex flex-col items-center p-12 space-y-6">
+      { !match 
+        ? 
+          <div className="loading loading-spinner"/>
+        : <>
+          <div className="card w-[80%] bg-base-200">
+            <div className="card-body">
+              <div className="card-title">
+                Match {match.number}
+              </div>
+              <form className="flex flex-col" onSubmit={submit} onChange={updateReport} >
+                { 
+                  ["Whole Match", ...match.blueAlliance, ...match.redAlliance].map((team, index) => (
+                    <div key={index}>
+                      <div className="label">
+                        <span className={`label-text text-lg ${typeof(team) === "number" 
+                          && (match.blueAlliance.includes(team) ? "text-blue-500" : "text-red-500")}`}>
+                          {team}
+                        </span>
+                      </div>
+                      <textarea id={`comment-${team.toString().replace(" ", "")}`} className="input input-bordered w-full h-20" placeholder="Enter comments here..."/>
+                    </div>
+                  ))
+                }
+                <button ref={submitButtonRef} className={`btn btn-${submitting ? "disabled" : "primary"} mt-6`}>
+                  {submitting ? <Loading bg="" size={42} /> : "Submit"}
+                </button>
+              </form>
+            </div>
+          </div>
+        </>
+      }
+    </div>
   );
 }
